requirements.txt
gradio==6.0.1 
numpy==2.3.5 
pillow==12.0.0 
torch==2.9.1+cpu 
torchvision==0.24.1+cpu 

